#3 feet: DDD 디렉토리 구조로 변경
참고 아티클 : https://cheese10yun.github.io/spring-guide-directory/
