--- src/main/kotlin/com/bside/v8/auth/adapter/input/controller/RegisterController.kt
@@ -1,24 +1,0 @@
-package com.bside.v8.auth.adapter.input.controller
-
-import com.bside.v8.auth.adapter.input.request.RegisterRequest
-import com.bside.v8.auth.adapter.input.response.AuthenticationResponse
-import com.bside.v8.auth.application.port.input.RegisterUseCase
-import com.bside.v8.global.annotation.WebAdapter
-import com.bside.v8.global.response.ApiResponseDto
-import org.springframework.web.bind.annotation.PostMapping
-import org.springframework.web.bind.annotation.RequestBody
-
-@WebAdapter
-class RegisterController(
-    private val registerUseCase: RegisterUseCase
-) {
-
-    @PostMapping("/api/v1/auth/register")
-    fun register(
-        @RequestBody registerRequest: RegisterRequest
-    ): ApiResponseDto<AuthenticationResponse> {
-        val token = registerUseCase.register(registerRequest.command())
-        val response = AuthenticationResponse(token)
-        return ApiResponseDto.OK(response)
-    }
-}
