--- src/main/kotlin/com/bside/v8/auth/adapter/input/response/AuthenticationResponse.kt
@@ -1,5 +1,0 @@
-package com.bside.v8.auth.adapter.input.response
-
-data class AuthenticationResponse(
-    val token: String
-)
