--- src/main/kotlin/com/bside/v8/auth/application/port/input/AuthenticationCommand.kt
@@ -1,6 +1,0 @@
-package com.bside.v8.auth.application.port.input
-
-data class AuthenticationCommand(
-    val email: String,
-    val password: String
-)
