--- src/main/kotlin/com/bside/v8/auth/application/port/input/AuthenticationUseCase.kt
@@ -1,6 +1,0 @@
-package com.bside.v8.auth.application.port.input
-
-interface AuthenticationUseCase {
-
-    fun authentication(command: AuthenticationCommand): String
-}
