--- src/main/kotlin/com/bside/v8/auth/application/port/input/RegisterUseCase.kt
@@ -1,6 +1,0 @@
-package com.bside.v8.auth.application.port.input
-
-interface RegisterUseCase {
-
-    fun register(command: RegisterCommand): String
-}
