--- src/main/kotlin/com/bside/v8/auth/application/port/output/FindAndCreateTokenPort.kt
@@ -1,6 +1,0 @@
-package com.bside.v8.auth.application.port.output
-
-interface FindAndCreateTokenPort {
-
-    fun findAndCreate(email: String): String
-}
