+++ src/main/kotlin/com/bside/v8/domain/member/application/SignUpService.kt
@@ -1,0 +1,14 @@
+package com.bside.v8.domain.member.application
+
+import com.bside.v8.domain.member.dao.persistence.SignUpPersistenceAdapter
+import com.bside.v8.domain.member.dto.command.RegisterCommand
+import org.springframework.stereotype.Service
+
+@Service
+class SignUpService(
+    private val signUpPersistenceAdapter: SignUpPersistenceAdapter
+) {
+
+    fun register(command: RegisterCommand): String =
+        signUpPersistenceAdapter.signUp(command.newUser())
+}
