+++ src/main/kotlin/com/bside/v8/domain/member/dao/repository/MemberRepository.kt
@@ -1,0 +1,12 @@
+package com.bside.v8.domain.member.dao.repository
+
+import com.bside.v8.domain.member.domain.Member
+import org.springframework.data.jpa.repository.JpaRepository
+import org.springframework.stereotype.Repository
+import java.util.Optional
+
+@Repository
+interface MemberRepository : JpaRepository<Member, Long> {
+
+    fun findByEmail(email: String): Optional<Member>
+}
