+++ src/main/kotlin/com/bside/v8/domain/member/dto/command/AuthenticationCommand.kt
@@ -1,0 +1,6 @@
+package com.bside.v8.domain.member.dto.command
+
+data class AuthenticationCommand(
+    val email: String,
+    val password: String
+)
