+++ src/main/kotlin/com/bside/v8/domain/member/dto/command/RegisterCommand.kt
@@ -1,0 +1,14 @@
+package com.bside.v8.domain.member.dto.command
+
+import com.bside.v8.domain.member.dto.domain.MemberDto
+import com.bside.v8.domain.model.Platform
+
+data class RegisterCommand(
+    val nickName: String,
+    val email: String,
+    val password: String,
+    val platform: Platform
+) {
+
+    fun newUser(): MemberDto = MemberDto(nickName, email, password, platform)
+}
