+++ src/main/kotlin/com/bside/v8/domain/member/dto/response/AuthenticationResponse.kt
@@ -1,0 +1,5 @@
+package com.bside.v8.domain.member.dto.response
+
+data class AuthenticationResponse(
+    val token: String
+)
