+++ src/main/kotlin/com/bside/v8/domain/model/Platform.kt
@@ -1,0 +1,8 @@
+package com.bside.v8.domain.model
+
+enum class Platform(val desc: String) {
+    EMAIL("이메일"),
+    GOOGLE("구글"),
+    APPLE("애플"),
+    KAKAO("카카오")
+}
