--- src/main/kotlin/com/bside/v8/user/domain/enumerate/Platform.kt
@@ -1,8 +1,0 @@
-package com.bside.v8.user.domain.enumerate
-
-enum class Platform(val desc: String) {
-    EMAIL("이메일"),
-    GOOGLE("구글"),
-    APPLE("애플"),
-    KAKAO("카카오")
-}
